refactor: move logger adapter
