+++ packages/core/src/logger.ts
@@ -1,6 +1,5 @@
 import { bgGreen, bgRed, bgBlue, bgYellow } from 'chalk'
 
-import { ILogger } from '@kondah/energizor'
 import { Label } from './label'
 
 export class Logger {
@@ -18,27 +17,5 @@ export class Logger {
 
   public warning(msg: string, label = 'WARNING'): void {
     console.log(bgYellow.bold.black(Label.create(label).value) + ' ' + msg)
-  }
-}
-
-export class EnergizorLoggerAdapter implements ILogger {
-  private readonly _label = Label.create('ENERGIZOR').value
-
-  public constructor(private readonly _logger: Logger) {}
-
-  public info(msg: string): void {
-    this._logger.info(msg, this._label)
-  }
-
-  public warning(msg: string): void {
-    this._logger.warning(msg, this._label)
-  }
-
-  public success(msg: string): void {
-    this._logger.success(msg, this._label)
-  }
-
-  public danger(msg: string): void {
-    this._logger.danger(msg, this._label)
   }
 }
